added eslint back, made ssr work properly
