+++ packages/frontend-next/.eslintignore
@@ -1,0 +1,3 @@
+package.json
+package-lock.json
+*.test.js
