+++ packages/frontend-next/package.json
@@ -6,18 +6,23 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
+    "lint": "eslint ./",
+    "lint-fix": "yarn lint -- --fix",
     "deploy": "../../node_modules/.bin/now -A ./frontend-next.json",
     "deploy-prod": "../../node_modules/.bin/now -A ./frontend-next.json --prod",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook"
   },
   "dependencies": {
+    "@apollo/react-components": "^3.1.3",
+    "@apollo/react-ssr": "^3.1.3",
+    "apollo-boost": "^0.4.7",
     "apollo-cache": "^1.3.2",
-    "apollo-cache-inmemory": "^1.6.5",
-    "apollo-client": "^2.6.8",
-    "apollo-link": "^1.2.13",
-    "apollo-link-batch-http": "^1.2.13",
+    "apollo-client": "^2.6.4",
+    "apollo-link": "^1.2.12",
     "apollo-utilities": "^1.3.2",
+    "eslint": "^6.8.0",
+    "eslint-plugin-react": "^7.17.0",
     "graphql": "^14.5.8",
     "graphql-tag": "^2.10.1",
     "mobile-detect": "^1.4.4",
