+++ packages/frontend-next/src/components/AppWrapper/appwrapper.stories.js
@@ -1,18 +1,18 @@
-import React from 'react';
-import AppWrapper from './appwrapper.component';
-import { StaticRouter } from 'react-router-dom';
-import Root from '../Root/root.component';
+import React from 'react'
+import AppWrapper from './appwrapper.component'
+import { StaticRouter } from 'react-router-dom'
+import Root from '../Root/root.component'
 
 export default {
-  title: 'AppWrapper',
-};
+	title: 'AppWrapper',
+}
 
 export const defaultStory = () => (
-  <StaticRouter location="/" context={{}}>
-    <Root>{() => <AppWrapper>content</AppWrapper>}</Root>
-  </StaticRouter>
-);
+	<StaticRouter location="/" context={{}}>
+		<Root>{() => <AppWrapper>content</AppWrapper>}</Root>
+	</StaticRouter>
+)
 
 defaultStory.story = {
-  name: 'default',
-};
+	name: 'default',
+}
