+++ packages/frontend-next/src/components/Button/button.component.js
@@ -1,3 +1,4 @@
+import React from 'react'
 import styled from 'styled-components'
 
 const Button = styled.button`
