+++ packages/frontend-next/src/components/DesktopNav/desktopnav.component.js
@@ -1,3 +1,4 @@
+import React from 'react'
 import styled from 'styled-components'
 import Text from '../Text/text.component'
