+++ packages/frontend-next/src/components/DesktopNav/desktopnav.stories.js
@@ -1,18 +1,18 @@
-import React from 'react';
-import DesktopNav from './desktopnav.component';
-import { StaticRouter } from 'react-router-dom';
-import Root from '../Root/root.component';
+import React from 'react'
+import DesktopNav from './desktopnav.component'
+import { StaticRouter } from 'react-router-dom'
+import Root from '../Root/root.component'
 
 export default {
-  title: 'DesktopNav',
-};
+	title: 'DesktopNav',
+}
 
 export const defaultStory = () => (
-  <StaticRouter location="/" context={{}}>
-    <Root>{() => <DesktopNav />}</Root>
-  </StaticRouter>
-);
+	<StaticRouter location="/" context={{}}>
+		<Root>{() => <DesktopNav />}</Root>
+	</StaticRouter>
+)
 
 defaultStory.story = {
-  name: 'default',
-};
+	name: 'default',
+}
