+++ packages/frontend-next/src/components/Entry/entry.component.js
@@ -1,3 +1,4 @@
+import React from 'react'
 import styled from 'styled-components'
 import Spacer from '../Spacer/spacer.component'
