+++ packages/frontend-next/src/components/Entry/entry.stories.js
@@ -1,18 +1,18 @@
-import React from 'react';
-import { StaticRouter } from 'react-router-dom';
-import Root from '../Root/root.component';
-import Entry from './entry.component';
+import React from 'react'
+import { StaticRouter } from 'react-router-dom'
+import Root from '../Root/root.component'
+import Entry from './entry.component'
 
 export default {
-  title: 'Entry',
-};
+	title: 'Entry',
+}
 
 export const defaultStory = () => (
-  <StaticRouter location="/" context={{}}>
-    <Root>{() => <Entry title="title" link="https://google.com" date="date" />}</Root>
-  </StaticRouter>
-);
+	<StaticRouter location="/" context={{}}>
+		<Root>{() => <Entry title="title" link="https://google.com" date="date" />}</Root>
+	</StaticRouter>
+)
 
 defaultStory.story = {
-  name: 'default',
-};
+	name: 'default',
+}
