+++ packages/frontend-next/src/components/Header/header.component.js
@@ -1,3 +1,4 @@
+import React from 'react'
 import styled from 'styled-components'
 import Icon from '../Icon/icon.component'
 import Grid from '../Grid/grid.component'
