+++ packages/frontend-next/src/components/Heart/heart.component.js
@@ -1,3 +1,5 @@
+
+import React from 'react'
 import Icon from '../../components/Icon/icon.component'
 import styled from 'styled-components'
