+++ packages/frontend-next/src/components/Heart/heart.stories.js
@@ -1,18 +1,18 @@
-import React from 'react';
-import { StaticRouter } from 'react-router-dom';
-import Root from '../Root/root.component';
-import Heart from './heart.component';
+import React from 'react'
+import { StaticRouter } from 'react-router-dom'
+import Root from '../Root/root.component'
+import Heart from './heart.component'
 
 export default {
-  title: 'Heart',
-};
+	title: 'Heart',
+}
 
 export const defaultStory = () => (
-  <StaticRouter location="/" context={{}}>
-    <Root>{() => <Heart />}</Root>
-  </StaticRouter>
-);
+	<StaticRouter location="/" context={{}}>
+		<Root>{() => <Heart />}</Root>
+	</StaticRouter>
+)
 
 defaultStory.story = {
-  name: 'default',
-};
+	name: 'default',
+}
