+++ packages/frontend-next/src/components/Loader/loader.stories.js
@@ -1,18 +1,18 @@
-import React from 'react';
-import { StaticRouter } from 'react-router-dom';
-import Root from '../Root/root.component';
-import Loader from './loader.component';
+import React from 'react'
+import { StaticRouter } from 'react-router-dom'
+import Root from '../Root/root.component'
+import Loader from './loader.component'
 
 export default {
-  title: 'Loader',
-};
+	title: 'Loader',
+}
 
 export const defaultStory = () => (
-  <StaticRouter location="/" context={{}}>
-    <Root>{() => <Loader />}</Root>
-  </StaticRouter>
-);
+	<StaticRouter location="/" context={{}}>
+		<Root>{() => <Loader />}</Root>
+	</StaticRouter>
+)
 
 defaultStory.story = {
-  name: 'default',
-};
+	name: 'default',
+}
