+++ packages/frontend-next/src/components/NoMatch/nomatch.stories.js
@@ -1,18 +1,18 @@
-import React from 'react';
-import { StaticRouter } from 'react-router-dom';
-import Root from '../Root/root.component';
-import NoMatch from './nomatch.component';
+import React from 'react'
+import { StaticRouter } from 'react-router-dom'
+import Root from '../Root/root.component'
+import NoMatch from './nomatch.component'
 
 export default {
-  title: 'NoMatch',
-};
+	title: 'NoMatch',
+}
 
 export const defaultStory = () => (
-  <StaticRouter location="/" context={{}}>
-    <Root>{() => <NoMatch />}</Root>
-  </StaticRouter>
-);
+	<StaticRouter location="/" context={{}}>
+		<Root>{() => <NoMatch />}</Root>
+	</StaticRouter>
+)
 
 defaultStory.story = {
-  name: 'default',
-};
+	name: 'default',
+}
