+++ packages/frontend-next/src/components/ResizeLoader/resizeloader.component.js
@@ -1,3 +1,4 @@
+import React from 'react'
 import styled from 'styled-components'
 
 const BouncyDiv = styled.div`
