+++ packages/frontend-next/src/components/ResizeLoader/resizeloader.stories.js
@@ -1,18 +1,18 @@
-import React from 'react';
-import { StaticRouter } from 'react-router-dom';
-import Root from '../Root/root.component';
-import ResizeLoader from './resizeloader.component';
+import React from 'react'
+import { StaticRouter } from 'react-router-dom'
+import Root from '../Root/root.component'
+import ResizeLoader from './resizeloader.component'
 
 export default {
-  title: 'ResizeLoader',
-};
+	title: 'ResizeLoader',
+}
 
 export const defaultStory = () => (
-  <StaticRouter location="/" context={{}}>
-    <Root>{() => <ResizeLoader />}</Root>
-  </StaticRouter>
-);
+	<StaticRouter location="/" context={{}}>
+		<Root>{() => <ResizeLoader />}</Root>
+	</StaticRouter>
+)
 
 defaultStory.story = {
-  name: 'default',
-};
+	name: 'default',
+}
