+++ packages/frontend-next/src/components/Text/text.component.js
@@ -1,3 +1,4 @@
+import React from 'react'
 import styled from 'styled-components'
 import { Link, Route } from 'react-router-dom'
