+++ packages/frontend-next/src/containers/RouterInfo/routerinfo.container.js
@@ -1,3 +1,4 @@
+import React from 'react'
 import { withRouter } from 'react-router'
 
 const RouterInfo = ({ context }) => {
