+++ packages/frontend-next/src/index.test.js
@@ -1,10 +1,10 @@
-import { shallow, mount } from 'enzyme';
-import App from './index';
+import { shallow, mount } from 'enzyme'
+import App from './index'
 
 it('renders without crashing shallow', () => {
-  shallow(<App />);
-});
+	shallow(<App />)
+})
 
 it('renders without crashing mount', () => {
-  mount(<App />).unmount();
-});
+	mount(<App />).unmount()
+})
